requirements.txt
asgiref==3.8.1
braintree==4.35.0
certifi==2025.4.26
charset-normalizer==3.4.2
dataclasses==0.6
Django==4.2
django-adminlte3==0.1.6
django-ckeditor==6.7.2
django-colorfield==0.14.0
django-js-asset==3.1.2
django-resized==1.0.3
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
idna==3.10
payme-pkg==2.6.5
pillow==11.2.1
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
